% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AncCond.R
\name{AncCond}
\alias{AncCond}
\title{Ancestral Condition Test}
\usage{
AncCond(
  tree,
  data,
  drop.state = NULL,
  model = "ARD",
  pi = "estimated",
  n.tails = 1,
  nsim = 100,
  iter = 100,
  message = F
)
}
\arguments{
\item{data}{This is the data matrix created by running \code{dataMatrx}}
}
\value{
constrained likelihood function is returned
}
\description{
Tests whether transitions in a discrete trait are associated with extreme
values of a continuous trait.
}
\details{
If the rate identity matrix is returned rates are indicated by the numbers:
\cr rate13 demipolyploidy for state2 - odd \cr \cr \cr The argument
\code{constrain} can contain any of the items listed below.  The default
state is shown. \cr \cr \code{drop.poly=F} \cr Sets polyploidy rate to zero
\cr
}
\note{
%% ~~further notes~~
}
\examples{

}
\references{
%% ~put references to the literature/web site here ~
}
\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\author{
Heath Blackmon
}
